requirements.txt
# Core dependencies for gemini-flow + THE_ORCHESTRATOR integration

# Google AI SDK
google-generativeai>=0.3.0

# Anthropic API (for THE_ORCHESTRATOR)
anthropic>=0.18.0

# Web framework (for Python API endpoints)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Async support
aiofiles>=23.2.1
aiohttp>=3.9.0

# Data handling
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
python-json-logger>=2.0.7

# THE_ORCHESTRATOR specific dependencies
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0  # For evolutionary algorithms
networkx>=3.2  # For agent graph structures

# Optional: For advanced features
redis>=5.0.0  # For distributed consciousness substrate
celery>=5.3.0  # For background task processing
psycopg2-binary>=2.9.0  # For PostgreSQL support

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0